requirement.txt
fastapi
uvicorn[standard]
pydantic 
jinja2

line-bot-sdk

firebase-admin